+++ CHANGES.rst
@@ -39,6 +39,7 @@ Enhancements
 Maintenance
 ------------
 
+* Add unit tests for shutters.
 * Set ``kind`` attribute to add plugin to ad.read_attrs list.
 * Clear ``PVPositionerSoftDone``'s setpoint & readback subscriptions at exit.